Make sure that the field sizes are odd
If the user's field base is an odd number, make the field's x and y axis odd by
reducing the fld_x and fld_y values.

Not a totally arbitrary decision to make them odd by substracting. If the user
had learned that the default y value is 9 by counting the rows in some output,
they could enter '-y 9' and actually get 9.

Seems like this whole situation could be improved...